Work on sound.
git-svn-id: https://cvet.by:8443/svn/FOnlineSource/Main@1312 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: daeb896c546bea722cf5295c3fd1364f6807310a

+++ Source/ItemHex.cpp
@@ -147,7 +147,7 @@ void ItemHex::Process()
     // Effect
     if( IsDynamicEffect() && !IsFinishing() )
     {
-        float dt = Timer::GameTick() - effLastTick;
+        float dt = (float) ( Timer::GameTick() - effLastTick );
         if( dt > 0 )
         {
             float speed = GetFlyEffectSpeed();
